length is decimal

+++ app/boats/models.py
@@ -24,7 +24,7 @@ class Boat(models.Model):
     name = models.CharField(max_length=50, unique=True)
     slug = AutoSlugField(populate_from='name')
 
-    length = models.PositiveSmallIntegerField(blank=True, null=True)
+    length = models.DecimalField(max_digits=3, decimal_places=1, blank=True, null=True)
     favorite = models.BooleanField(default=False)
 
     bw_url = models.URLField(blank=True)
